+++ Source/RadiantPi.Kaleidescape/IKaleidescape.cs
@@ -1,6 +1,6 @@
 /*
  * RadiantPi.Kaleidescape - Communication client for Kaleidescape
- * Copyright (C) 2020-2022 - Steve G. Bjorg
+ * Copyright (C) 2020-2023 - Steve G. Bjorg
  *
  * This program is free software: you can redistribute it and/or modify it
  * under the terms of the GNU Affero General Public License as published by the
@@ -25,6 +25,8 @@ public interface IKaleidescape : IDisposable {
 
     //--- Events ---
     event EventHandler<HighlightedSelectionChangedEventArgs>? HighlightedSelectionChanged;
+    event EventHandler<UiStateChangedEventArgs>? UiStateChanged;
+    event EventHandler<MovieLocationEventArgs>? MovieLocationChanged;
 
     //--- Methods ---
     Task ConnectAsync();