+++ ReleaseNotes.md
@@ -1,5 +1,28 @@
 # Release Notes
 
+## v2.2 (2023-02-14)
+
+### Features
+
+* Added `MovieLocationChanged` event triggered by MOVIE_LOCATION notification.
+* Added 10 seconds timeout to `GetContentDetailsAsync()` operation.
+
+
+## v2.1 (2022-07-12)
+
+### Features
+
+* Added `UiStateChanged` event triggered by UI_STATE notifications.
+
+### Samples
+
+* Renamed _ShowSelection_ to _ShowEvents_
+    * Display UI state changed events
+
+### Misc.
+
+* Upgraded _Microsoft.Extensions.Logging_ to 6.0.0
+
 
 ## v2.0 (2022-02-15)